Move file upload to separate component.
