+++ src/app/[locale]/klage/opprett/[id]/[vedtakId]/consts.ts
@@ -1,0 +1,4 @@
+export const MAX_LEN_BACKGROUND = 1000;
+export const MAX_FILES = 3;
+export const MAX_SIZE_MB = 1;
+export const MAX_SIZE = MAX_SIZE_MB * 1024 * 1024;
