+++ src/app/[locale]/klage/opprett/[id]/[vedtakId]/klageForm.tsx
@@ -1,18 +1,8 @@
 "use client";
 
-import {
-    Alert,
-    Button,
-    FileObject,
-    FileRejected,
-    FileRejectionReason,
-    FileUpload,
-    Heading,
-    Textarea,
-    VStack,
-} from "@navikt/ds-react";
+import { Alert, Button, FileObject, Textarea } from "@navikt/ds-react";
 import { useTranslations } from "next-intl";
-import { useForm, SubmitHandler, useController, Resolver } from "react-hook-form";
+import { useForm, SubmitHandler, Resolver } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import z from "zod";
 import { useQueryClient } from "@tanstack/react-query";
@@ -27,12 +17,10 @@ import {
 } from "../../../../../../generated/klage-controller/klage-controller";
 import useVedtakId from "../../../../../../hooks/useVedtakId";
 
-const MAX_LEN_BACKGROUND = 1000;
-const MAX_FILES = 3;
-const MAX_SIZE_MB = 1;
-const MAX_SIZE = MAX_SIZE_MB * 1024 * 1024;
+import FileUploadField from "./FileUploadField";
+import { MAX_LEN_BACKGROUND, MAX_FILES } from "./consts";
 
-type FormValues = {
+export type FormValues = {
     background: string | null;
     files?: FileObject[];
 };
@@ -85,25 +73,6 @@ const KlageForm = () => {
         }
     };
 
-    // Integrate FileUpload with react-hook-form
-    const {
-        field: { value: files = [], onChange },
-    } = useController({
-        name: "files",
-        control,
-    });
-
-    const onFileSelect = (newFiles: FileObject[]) => {
-        onChange([...files, ...newFiles]);
-    };
-
-    const removeFile = (fileToRemove: FileObject) => {
-        onChange(files.filter((file: FileObject) => file !== fileToRemove));
-    };
-
-    const acceptedFiles = files.filter((file) => !file.error);
-    const rejectedFiles = files.filter((f): f is FileRejected => f.error);
-
     return (
         <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-20">
             <Textarea
@@ -112,57 +81,7 @@ const KlageForm = () => {
                 error={errors.background?.message && t(errors.background.message)}
                 {...register("background")}
             ></Textarea>
-            <VStack gap="6">
-                <FileUpload.Dropzone
-                    label="Last opp filer til søknaden"
-                    description={`Du kan laste opp Word- og PDF-filer. Maks 3 filer. Maks størrelse ${MAX_SIZE_MB} MB.`}
-                    accept=".doc,.docx,.pdf"
-                    maxSizeInBytes={MAX_SIZE}
-                    fileLimit={{ max: MAX_FILES, current: acceptedFiles.length }}
-                    onSelect={(newFiles) => onFileSelect(newFiles)}
-                />
-                {acceptedFiles.length > 0 && (
-                    <VStack gap="2">
-                        <Heading level="3" size="xsmall">
-                            {`Vedlegg (${acceptedFiles.length})`}
-                        </Heading>
-                        <VStack as="ul" gap="3">
-                            {acceptedFiles.map((file, index) => (
-                                <FileUpload.Item
-                                    as="li"
-                                    key={index}
-                                    file={file.file}
-                                    button={{
-                                        action: "delete",
-                                        onClick: () => removeFile(file),
-                                    }}
-                                />
-                            ))}
-                        </VStack>
-                    </VStack>
-                )}
-                {rejectedFiles.length > 0 && (
-                    <VStack gap="2">
-                        <Heading level="3" size="xsmall">
-                            Vedlegg med feil
-                        </Heading>
-                        <VStack as="ul" gap="3">
-                            {rejectedFiles.map((rejected, index) => (
-                                <FileUpload.Item
-                                    as="li"
-                                    key={index}
-                                    file={rejected.file}
-                                    error={fileErrors[rejected.reasons[0] as FileRejectionReason]}
-                                    button={{
-                                        action: "delete",
-                                        onClick: () => removeFile(rejected),
-                                    }}
-                                />
-                            ))}
-                        </VStack>
-                    </VStack>
-                )}
-            </VStack>
+            <FileUploadField name="files" control={control} />
             <Button loading={lastOppVedleggMutation.isPending || sendKlageMutation.isPending} type="submit">
                 Send klage
             </Button>
@@ -171,11 +90,6 @@ const KlageForm = () => {
             )}
         </form>
     );
-};
-
-const fileErrors: Record<FileRejectionReason, string> = {
-    fileType: "Filformatet støttes ikke",
-    fileSize: `Filen er større enn ${MAX_SIZE_MB} MB`,
 };
 
 export default KlageForm;
